+++ app/src/main/res/layout/fragment_share_badge.xml
@@ -13,54 +13,57 @@
     <FrameLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-
             tools:context=".ui.share.ShareBadgeFragment">
+
         <RelativeLayout
-                android:layout_width="match_parent"
+                android:id="@+id/capture_layout_badge"
+                android:layout_width="264dp"
+                android:layout_height="368dp"
                 android:layout_marginLeft="@dimen/pageMarginAndOffset"
                 android:layout_marginRight="@dimen/pageMarginAndOffset"
-                android:layout_height="match_parent"
                 android:background="@drawable/sharebackground_level01"
                 android:orientation="vertical">
-        <androidx.constraintlayout.widget.ConstraintLayout
-                android:layout_width="match_parent"
-                android:layout_height="match_parent">
 
-            <LinearLayout
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:layout_marginTop="64dp"
-                    android:gravity="center"
-                    android:orientation="vertical"
-                    app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent">
+            <androidx.constraintlayout.widget.ConstraintLayout
 
-                <TextView
+                    android:layout_width="match_parent"
+                    android:layout_height="match_parent">
+
+                <LinearLayout
                         android:layout_width="wrap_content"
                         android:layout_height="wrap_content"
-                        android:text="34"
-                        android:textColor="@color/primary_white"
-                        android:textSize="80sp"
+                        android:layout_marginTop="64dp"
+                        android:gravity="center"
+                        android:orientation="vertical"
                         app:layout_constraintEnd_toEndOf="parent"
                         app:layout_constraintStart_toStartOf="parent"
-                        app:layout_constraintTop_toTopOf="parent" />
+                        app:layout_constraintTop_toTopOf="parent">
 
-                <TextView
-                        android:layout_width="wrap_content"
-                        android:layout_height="24dp"
-                        android:layout_marginTop="8dp"
-                        android:text="분"
-                        android:textColor="@color/primary_white"
-                        android:textSize="17sp"
-                        android:textStyle="bold"
-                        app:layout_constraintEnd_toEndOf="parent"
-                        app:layout_constraintStart_toStartOf="parent"
-                        app:layout_constraintTop_toTopOf="parent" />
+                    <TextView
+                            android:layout_width="wrap_content"
+                            android:layout_height="wrap_content"
+                            android:text="34"
+                            android:textColor="@color/primary_white"
+                            android:textSize="80sp"
+                            app:layout_constraintEnd_toEndOf="parent"
+                            app:layout_constraintStart_toStartOf="parent"
+                            app:layout_constraintTop_toTopOf="parent" />
 
-            </LinearLayout>
+                    <TextView
+                            android:layout_width="wrap_content"
+                            android:layout_height="24dp"
+                            android:layout_marginTop="8dp"
+                            android:text="분"
+                            android:textColor="@color/primary_white"
+                            android:textSize="17sp"
+                            android:textStyle="bold"
+                            app:layout_constraintEnd_toEndOf="parent"
+                            app:layout_constraintStart_toStartOf="parent"
+                            app:layout_constraintTop_toTopOf="parent" />
 
-        </androidx.constraintlayout.widget.ConstraintLayout>
+                </LinearLayout>
+
+            </androidx.constraintlayout.widget.ConstraintLayout>
         </RelativeLayout>
     </FrameLayout>
 </layout>
